requirements.txt
openai==1.97.1
psycopg2==2.9.10
requests==2.32.4
google-genai==1.26.0